+++ packages/bpk-docs/src/pages/BannerAlertsPage/BannerAlertsPage.js
@@ -16,13 +16,19 @@
  * limitations under the License.
  */
 
-import React from 'react';
+import React, { Component } from 'react';
+import { fontWeightBold } from 'bpk-tokens/tokens/base.es6';
+import PropTypes from 'prop-types';
+import { cssModules } from 'bpk-react-utils';
 import BpkBannerAlert, { ALERT_TYPES } from 'bpk-component-banner-alert';
-
 import bannerAlertReadme from 'bpk-component-banner-alert/readme.md';
-
 import DocsPageBuilder from './../../components/DocsPageBuilder';
 import Paragraph from './../../components/Paragraph';
+
+import STYLES from './bpk-banner-alerts-page.scss';
+
+const getClassName = cssModules(STYLES);
+const componentClassName = getClassName('bpk-banner-alerts-page__component');
 
 const longMessage = `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque sagittis sagittis purus, id
 blandit ipsum. Pellentesque nec diam nec erat condimentum dapibus. Nunc diam augue, egestas id egestas ut, facilisis
@@ -30,7 +36,49 @@ nec mi. Donec et congue odio, nec laoreet est. Integer rhoncus varius arcu, a fr
 porta varius ullamcorper. Sed laoreet libero mauris, non pretium lectus accumsan et. Suspendisse vehicula ullamcorper
 sapien, et dapibus mi aliquet non. Pellentesque auctor sagittis lectus vitae rhoncus. Fusce id enim porttitor, mattis
 ante in, vestibulum nulla.`;
-const richMessage = <span style={{ fontWeight: 700 }}>Successful alert with custom rendered message</span>;
+const richMessage = <span style={{ fontWeight: fontWeightBold }}>Successful alert with custom rendered message</span>;
+
+class BpkBannerDismissable extends Component {
+  constructor() {
+    super();
+
+    this.setDismissed = this.setDismissed.bind(this);
+
+    this.state = {
+      show: true,
+    };
+  }
+
+  setDismissed() {
+    this.setState({
+      show: false,
+    });
+  }
+
+  render() {
+    return (
+      <BpkBannerAlert
+        className={componentClassName}
+        message={this.props.message}
+        type={this.props.type}
+        dismissable
+        onDismiss={this.setDismissed}
+        show={this.state.show}
+        dismissButtonLabel="Dismiss"
+      />
+    );
+  }
+}
+
+BpkBannerDismissable.propTypes = {
+  message: PropTypes.string,
+  type: PropTypes.string,
+};
+
+BpkBannerDismissable.defaultProps = {
+  message: null,
+  type: null,
+};
 
 const components = [
   {
@@ -42,18 +90,31 @@ const components = [
       </Paragraph>,
     ],
     examples: [
-      <BpkBannerAlert message="Neutral alert." type={ALERT_TYPES.NEUTRAL} />,
-      <br />,
-      <BpkBannerAlert message="Successful alert." type={ALERT_TYPES.SUCCESS} />,
-      <br />,
       <BpkBannerAlert
+        className={componentClassName}
+        message="Neutral alert."
+        type={ALERT_TYPES.NEUTRAL}
+      />,
+      <BpkBannerAlert
+        className={componentClassName}
+        message="Successful alert."
+        type={ALERT_TYPES.SUCCESS}
+      />,
+      <BpkBannerAlert
+        className={componentClassName}
         message={richMessage}
         type={ALERT_TYPES.SUCCESS}
       />,
-      <br />,
-      <BpkBannerAlert message="Warn alert." type={ALERT_TYPES.WARN} />,
-      <br />,
-      <BpkBannerAlert message="Error alert." type={ALERT_TYPES.ERROR} />,
+      <BpkBannerAlert
+        className={componentClassName}
+        message="Warn alert."
+        type={ALERT_TYPES.WARN}
+      />,
+      <BpkBannerAlert
+        className={componentClassName}
+        message="Error alert."
+        type={ALERT_TYPES.ERROR}
+      />,
     ],
   },
   {
@@ -67,32 +128,76 @@ const components = [
     ],
     examples: [
       <BpkBannerAlert
+        className={componentClassName}
         message="Neutral alert with more information."
         type={ALERT_TYPES.NEUTRAL}
         toggleButtonLabel="See more"
       >
         {longMessage}
       </BpkBannerAlert>,
-      <br />,
       <BpkBannerAlert
+        className={componentClassName}
         message="Successful alert with more information."
         type={ALERT_TYPES.SUCCESS}
         toggleButtonLabel="See more"
       >
         {longMessage}
       </BpkBannerAlert>,
-      <br />,
-      <BpkBannerAlert message="Warn alert with more information." type={ALERT_TYPES.WARN} toggleButtonLabel="See more">
+      <BpkBannerAlert
+        className={componentClassName}
+        message="Warn alert with more information."
+        type={ALERT_TYPES.WARN}
+        toggleButtonLabel="See more"
+      >
         {longMessage}
       </BpkBannerAlert>,
-      <br />,
       <BpkBannerAlert
+        className={componentClassName}
         message="Error alert with more information."
         type={ALERT_TYPES.ERROR}
         toggleButtonLabel="See more"
       >
         {longMessage}
       </BpkBannerAlert>,
+    ],
+  },
+  {
+    id: 'dismissable',
+    title: 'Dismissable',
+    blurb: [
+      <Paragraph>
+        Banner alerts can be configured to include a close icon so that the user can dismiss them.
+      </Paragraph>,
+    ],
+    examples: [
+      <BpkBannerDismissable
+        className={componentClassName}
+        message="Neutral alert with dismiss option."
+        type={ALERT_TYPES.NEUTRAL}
+        dismissable
+        dismissButtonLabel="Dismiss"
+      />,
+      <BpkBannerDismissable
+        className={componentClassName}
+        message="Successful alert with dismiss option."
+        type={ALERT_TYPES.SUCCESS}
+        dismissable
+        dismissButtonLabel="Dismiss"
+      />,
+      <BpkBannerDismissable
+        className={componentClassName}
+        message="Warn alert with dismiss option."
+        type={ALERT_TYPES.WARN}
+        dismissable
+        dismissButtonLabel="Dismiss"
+      />,
+      <BpkBannerDismissable
+        className={componentClassName}
+        message="Error alert with dismiss option."
+        type={ALERT_TYPES.ERROR}
+        dismissable
+        dismissButtonLabel="Dismiss"
+      />,
     ],
   },
 ];